remove decodeParams.js, facebook.js, objectToParam.js. Added FaceBook.js
