+++ client/src/actions/index.js
@@ -1,6 +1,8 @@
 import axios from 'axios';
 import { FETCH_USER, FETCH_SURVEYS, FETCH_EVENTS } from './types';
 
+
+
 export const fetchUser = () => async dispatch => {
   const res = await axios.get('/api/current_user');
