+++ client/src/components/Dashboard/FaceBook.js
@@ -1,0 +1,77 @@
+import React, { Component } from "react";
+import FacebookLogin from 'react-facebook-login/dist/facebook-login-render-props';
+import styled from "styled-components";
+
+
+const Ul = styled.ul`
+  display: flex;
+  li {
+    align-self: center;
+    margin-left: 5px;
+  }
+
+  span {
+    color: ${props => props.theme.colors.secondary};
+    font-weight: 600;
+  }
+`;
+
+export default class Facebook extends Component {
+  state = {
+    isLoggedIn: false,
+    userID: "",
+    name: "",
+    email: "",
+    picture: ""
+  };
+
+  responseFacebook = response => {
+    // console.log(response);
+
+    this.setState({
+      isLoggedIn: true,
+      userID: response.userID,
+      name: response.name,
+      email: response.email,
+      picture: response.picture.data.url
+    });
+  };
+
+  componentClicked = () => console.log("clicked");
+
+  render() {
+    let fbContent;
+
+    if (this.state.isLoggedIn) {
+      fbContent = (
+        <div
+          style={{
+            width: "200px",
+            margin: "right",
+            padding: "5px"
+          }}
+        >
+          <img src={this.state.picture} alt={this.state.name} />
+          <h2>Welcome {this.state.name}</h2>
+          Email: {this.state.email}
+        </div>
+      );
+    } else {
+      fbContent = (
+        <FacebookLogin
+          appId="2401284699888073"
+          autoLoad={true}
+          fields="name,email,picture"
+          onClick={this.componentClicked}
+          callback={this.responseFacebook}
+          size="small"
+          render={renderProps => (
+            <button onClick={renderProps.onClick}>This is my custom FB button</button>
+          )}
+        />
+      );
+    }
+
+    return <div>{fbContent}</div>;
+  }
+}
